+++ ReleaseNotes/release_notes.md
@@ -16,6 +16,7 @@
  1. fit two small typos in `ostap/math/derivative.py` (Thanks to Dmitry Golubkov) 
  1. fix the issue with `BernsteinEven`
  1. fix call for `RooFormulaVar::formula` for old versions of ROOT 
+ 1. fix efficiency tests 
 
 # v1.6.4.0